increase test coverage and break API
util#send2box is now ftpsession#send2box to avoid circular imports
